Separate prop class into separate file
Makes code cleaner, easier to see what each file is for / does.
