+++ Assets/Scripts/Prop.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 48f0d53c3d28b414187a97438ab3e40c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
